1.主畫面新增「關閉所有視窗」、「登出」功能 2.修正登入時用不存在帳號會造成程式錯誤的問題 3.修正商品編輯舉面 新增資料，若指定商品類別不存在任何商   時，自動帶入商品ID功能錯誤的問題 PS:修正2請同時更新資料庫備份

+++ 進銷存系統/進銷存系統/SqlFun/SqlFunUser.cs
@@ -59,7 +59,7 @@ namespace 進銷存系統.SqlFun
 
             ObjectParameter objIsAdmin = new ObjectParameter("IsAdmin", typeof(byte));
             SQLData.db.使用者LoginByEntity(LoginID, LoginPW, objIsAdmin);
-            int IsAdmin = (byte)objIsAdmin.Value;
+            short IsAdmin = (short)objIsAdmin.Value;
             if (IsAdmin >= 0)
             {
                 bIsLogin = true;
